+++ lib/src/localization/languages/portuguese_brazillian_language.dart
@@ -54,6 +54,9 @@ class PortugueseBrasillianLanguage extends Language {
               "'{PropertyName}' deve ser maior ou igual à data '{StartValue}' e menor ou igual à data '{EndValue}'.",
           Language.code.exclusiveBetweenDatetime:
               "'{PropertyName}' deve ser maior que a data '{StartValue}' e menor que a data '{EndValue}'.",
+          Language.code.validPhoneBr: "'{PropertyName}' não é um telefone válido.",
+          Language.code.validPhoneDdiBr: "'{PropertyName}' não é um telefone DDI válido."
+              "'{PropertyName}' deve ser maior que a data '{StartValue}' e menor que a data '{EndValue}'.",
           Language.code.sequentialRepeatedCharacters:
               "'{PropertyName}' não pode ter caracteres repetidos em sequência.",
           Language.code.sequentialCharactersNotAllowed: