+++ lib/src/validations/validations.dart
@@ -30,5 +30,6 @@ part 'valid_cpf_validation.dart';
 part 'valid_creditcard_validation.dart';
 part 'valid_email_validation.dart';
 part 'valid_cpf_or_cnpj_validation.dart';
+part 'valid_phone_validation.dart';
 part 'repeated_characters_validation.dart';
 part 'sequential_characters_validation.dart';