+++ test/lucid_validation_test.dart
@@ -3,9 +3,7 @@ import 'package:test/test.dart';
 import 'mocks/mocks.dart';
 
 void main() {
-  test(
-      'when validating [UserEntityMock] should return a list of error messages for the email field',
-      () {
+  test('when validating [UserEntityMock] should return a list of error messages for the email field', () {
     final validator = UserValidator();
     final userEntity = UserModel()
       ..age = 18
@@ -16,12 +14,10 @@ void main() {
     final result = validator.validate(userEntity);
     final exceptions = result.exceptions;
 
-    expect(exceptions.length, 4);
+    expect(exceptions.length, 2);
   });
 
-  test(
-      'when validating [UserModel] should return a list of error messages for the password field',
-      () {
+  test('when validating [UserModel] should return a list of error messages for the password field', () {
     final validator = UserValidator();
     final userEntity = UserModel()
       ..age = 18
@@ -32,12 +28,10 @@ void main() {
     final result = validator.validate(userEntity);
     final exceptions = result.exceptions;
 
-    expect(exceptions.length, 8);
+    expect(exceptions.length, 6);
   });
 
-  test(
-      'when validating [UserModel] should return a list of error messages for the age field',
-      () {
+  test('when validating [UserModel] should return a list of error messages for the age field', () {
     final validator = UserValidator();
     final userEntity = UserModel()
       ..age = 15
@@ -48,12 +42,10 @@ void main() {
     final result = validator.validate(userEntity);
     final exceptions = result.exceptions;
 
-    expect(exceptions.length, 3);
+    expect(exceptions.length, 1);
   });
 
-  test(
-      'when validating [UserModel] should return a list of error messages for the phone field',
-      () {
+  test('when validating [UserModel] should return a list of error messages for the phone field', () {
     final validator = UserValidator();
 
     final userEntity = UserModel()
@@ -65,19 +57,17 @@ void main() {
     final result = validator.validate(userEntity);
     final exceptions = result.exceptions;
 
-    expect(exceptions.length, 3);
+    expect(exceptions.length, 1);
   });
 
-  test(
-      'when validating [UserModel] should return a list of error messages for all fields',
-      () {
+  test('when validating [UserModel] should return a list of error messages for all fields', () {
     final validator = UserValidator();
     final userEntity = UserModel()..age = 15;
 
     final result = validator.validate(userEntity);
     final exceptions = result.exceptions;
 
-    expect(exceptions.length, 12);
+    expect(exceptions.length, 10);
   });
 
   test('EqualTo', () {